+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 - [ ] 改Layout,先抄Element-plus個Docs, 參考vitepress, https://vite-starter.element-plus.org/
   - [x] App.vue
+  - [x] layouts/BaseSide.vue - icon-Meun < 未得 @21/10/2022
   - [ ] layouts/BaseHeader.vue - dark.js < 未得
-  - [ ] layouts/BaseSide.vue - icon-Meun < 未得
 
 - [ ] 製作測試用頁面
   - [ ] 入口在View建立Test.vue載入components